Minor fixes and refactoring - Rename Strings to Messages - Set activity log's modified_on within __init__ - Change ActivityLog constructor from (action, time, event, user) to (user, action, time, details) - Define is_registered_in_event in User model - Change event __repr__ to only show Start/End time of event - Return modification count on remove_car - Remove ensure_event_exist from addcar/rmcar - Add car to activity log detail of addcar/rmcar - Change la command to lsuser - Remove allowed_user from lshow

+++ src/plugins/lottery.py
@@ -3,13 +3,13 @@
 import re
 from mmpy_bot.bot import respond_to
 from mmpy_bot.utils import allow_only_direct_message, allowed_users
-from .utils import ensure_user_exist,ensure_event_exist
-from db.repository import UserRepository, EventRepository, ActivityLogRepository
-from db.models import ActivityLog
-from shared import State, Action
+from .utils import ensure_user_exist, ensure_event_exist
+from db.repository import UserRepository, EventRepository, \
+    ActivityLogRepository
+from db.models import ActivityLog, Car
+from shared import Action
 from settings.settings import ADMINS
-from .messages import Strings
-from datetime import datetime
+from .messages import Messages
 
 
 @respond_to(r'^reg\s*$', re.IGNORECASE)
@@ -17,17 +17,14 @@ from datetime import datetime
 @ensure_event_exist()
 @ensure_user_exist()
 def register(message, user, active_event):
-
     if user.car is None:
-        return message.send(Strings.DEFINE_CAR)
-    
-    state_for_active_event = list(filter(lambda c: c['event_id'] == active_event.event_id and c['state'] == State.REGISTERED.name, user.user_state))
+        return message.send(Messages.DEFINE_CAR)
 
-    if  (len(state_for_active_event) <1 ):
-        UserRepository().participate(user.user_id,active_event.event_id)
-        message.react(Strings.SUCCESSFUL_REGISTERATION)
+    if not user.is_registered_in_event(active_event.event_id):
+        UserRepository().participate(user.user_id, active_event.event_id)
+        message.react(Messages.SUCCESSFUL_REGISTERATION)
     else:
-        message.send(Strings.ALREADY_REGISTERED)
+        message.send(Messages.ALREADY_REGISTERED)
 
 
 @respond_to(r'^unreg\s*$', re.IGNORECASE)
@@ -35,58 +32,75 @@ def register(message, user, active_event):
 @ensure_event_exist()
 @ensure_user_exist()
 def withdraw(message, user, active_event):
-           
-    state_for_active_event = list(filter(lambda c: c['event_id'] == active_event.event_id and c['state'] == State.REGISTERED.name, user.user_state))
-
-    if len( state_for_active_event) > 0 :
-        UserRepository().withdraw(user.user_id,active_event.event_id)
-        message.send(Strings.UNDO_REGISTERATION)
+    if user.is_registered_in_event(active_event.event_id):
+        UserRepository().withdraw(user.user_id, active_event.event_id)
+        message.send(Messages.UNDO_REGISTERATION)
     else:
-        message.send(Strings.NOT_YET_REGISTERED)
+        message.send(Messages.NOT_YET_REGISTERED)
 
 
 @respond_to(r'^mycar\s*$', re.IGNORECASE)
 @allow_only_direct_message()
 @ensure_user_exist()
 def mycar(message, user):
-    message.send(user.car.__repr__() if user.car is not None else Strings.NO_CAR_EXISTS)
+    message.send(user.car.__repr__()
+                 if user.car is not None else Messages.NO_CAR_EXISTS)
 
 
 @respond_to(r'^rmcar\s*$', re.IGNORECASE)
 @allow_only_direct_message()
-@ensure_event_exist()
 @ensure_user_exist()
-def remove_car(message, user, active_event):
+def remove_car(message, user):
+    if user.car is None:
+        return message.send(Messages.NO_CAR_EXISTS)
+
     UserRepository().remove_car(user.user_id)
-    UserRepository().withdraw(user.user_id,active_event.event_id)
-    activity_log=ActivityLog(Action.RMCAR.name, datetime.now(),active_event.event_id,user.user_id )
+
+    activity_details = {"car": user.car}
+    activity_log = ActivityLog(user.user_id, Action.RMCAR.name,
+                               details=activity_details)
     ActivityLogRepository().log_action(activity_log)
-    message.send(Strings.CAR_REMOVED)
+
+    active_event = EventRepository().find_active_event()
+    if active_event is not None:
+        UserRepository().withdraw(user.user_id, active_event.event_id)
+        return message.send(Messages.CAR_REMOVED_AND_WITHDRAWED)
+
+    message.send(Messages.CAR_REMOVED)
 
 
 @respond_to(r'^addcar ([\w\s\d]+) - ((?:ایران|ايران|iran|ir)[\s]*[\d]{2} '
             r'[\d]{2}[\w]{1}[\d]{3})$', re.IGNORECASE)
 @allow_only_direct_message()
-@ensure_event_exist()
 @ensure_user_exist()
-def add_car(message, user, active_event, model, plate_number, *args, **kwargs):
+def add_car(message, user, model, plate_number, *args, **kwargs):
     UserRepository().add_car(user.user_id, model, plate_number)
-    activity_log=ActivityLog(Action.ADDCAR.name, datetime.now(),active_event.event_id,user.user_id )
-    ActivityLogRepository().log_action(activity_log)
-    UserRepository().participate(user.user_id,active_event.event_id)
-    message.send(Strings.CAR_REGISTERED)
 
-@respond_to(r'^ls\s*$', re.IGNORECASE)
+    activity_details = {"car": Car(plate_number, model)}
+    activity_log = ActivityLog(user.user_id, Action.ADDCAR.name,
+                               details=activity_details)
+    ActivityLogRepository().log_action(activity_log)
+
+    active_event = EventRepository().find_active_event()
+    if active_event is not None:
+        UserRepository().participate(user.user_id, active_event.event_id)
+        return message.send(Messages.CAR_AND_PARTICIPATION_REGISTERED)
+
+    message.send(Messages.CAR_REGISTERED)
+
+
+@respond_to(r'^lls\s*$', re.IGNORECASE)
 @allow_only_direct_message()
 @ensure_event_exist()
 @allowed_users(*ADMINS)
 def list_participants(message, active_event):
     users = UserRepository().find_participants(active_event.event_id)
-    usernames = '\n'.join(map(lambda u: "%s, %s" % (u.username, u.car.plate_number), users))
+    usernames = '\n'.join(map(lambda u: "%s, %s" %
+                              (u.username, u.car.plate_number), users))
     message.send(usernames)
 
 
-@respond_to(r'^la\s*$', re.IGNORECASE)
+@respond_to(r'^lsuser\s*$', re.IGNORECASE)
 @allow_only_direct_message()
 @allowed_users(*ADMINS)
 def list_users(message):
@@ -101,24 +115,24 @@ def list_users(message):
 def add_event(message, duration, unit):
     active_event = EventRepository().find_active_event()
     if active_event is None:
+        # ToDo move duration calculation to add_event method
         duration = int(duration)
         if unit == "d":
             duration = 24 * duration
         EventRepository().add_event(duration)
-        message.send(Strings.NEW_EVENT_REGISTERED)
+        message.send(Messages.NEW_EVENT_REGISTERED)
     else:
-        message.send(Strings.EVENT_ALREADY_EXISTS)
+        message.send(Messages.EVENT_ALREADY_EXISTS)
 
 
 @respond_to(r'^lshow\s*$', re.IGNORECASE)
 @allow_only_direct_message()
-@allowed_users(*ADMINS)
 def get_events(message):
     event = EventRepository().find_active_event()
-    if event is not None:      
-       message.send(event.__repr__())
+    if event is not None:
+        message.send(event.__repr__())
     else:
-        message.send(Strings.NOT_VALID_EVENT)
+        message.send(Messages.NOT_VALID_EVENT)
 
 
 @respond_to(r'^lclose\s*$', re.IGNORECASE)
@@ -127,6 +141,6 @@ def get_events(message):
 def delete_event(message):
     event = EventRepository().deactive_event()
     if event is True:
-        message.send(Strings.EVENT_DEACTIVED)
+        message.send(Messages.EVENT_DEACTIVED)
     else:
-        message.send(Strings.NOT_VALID_EVENT)
+        message.send(Messages.NOT_VALID_EVENT)
